requirements.txt
Flask
flask-restx
Flask-Login
Flask-SQLAlchemy
Flask-Migrate
flask-cors
Bootstrap-Flask
Flask-WTF
WTForms-SQLAlchemy
email_validator
flask-jwt-extended
gunicorn
pyyaml
python-dateutil
